requirements.txt
requests
beautifulsoup4
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
